feat: get id de routine, crear swagger de routine y crear validación extra en validate.ts. fix: borrar post [athlete] y modificar validaciones en el  patch
